Move License Information to README
Removes the existing LICENSE file which would otherwise be copied into any new projects using this template, even if those projects are proprietary. In this case, it's probably better to just reference that this project is licensed under the MIT license and link to the LICENSE file from the SaltLite framework repo.